+++ src/PipedEverything/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.18.0] U52-626616-SC (release)
+- removed default capacity; undefined storages stay unchanged
+- fixed building ceasing operation (from v1.0.3) if storage full while pipe disconnected
+
 ## [1.0.17.0] U52-626616-SC (release)
 - increased default liquid capacity to 400kg
 - increased AlgaeHabitat capacity to 800kg, fixing manual emptying